Refactoring WaveGraph to GraphModel. Creating new Crates WaveModel and WaveGraph, which will later be used. Creating Struct "WaveModel" inside "wavemodel" Beginning to write Documentation for WaveModel.
